requirements.txt
feedparser==6.0.8
requests==2.26.0
googletrans==3.1.0a0
python-telegram-bot==13.7
schedule==1.1.0
urllib3==1.26.6
pyTelegramBotAPI
python-dotenv
nltk
scikit-learn
beautifulsoup4